+++ edgelet/edgelet-core/Cargo.toml
@@ -20,6 +20,7 @@ serde_derive = "1.0"
 serde_json = "1.0"
 sha2 = "0.7.0"
 log = "0.4"
+parse_duration = "2.0.1"
 url = "1.7"
 url_serde = "0.2"
 tokio = "0.1"
