+++ edgelet/iotedge/Cargo.toml
@@ -19,7 +19,6 @@ hyper = "0.12"
 lazy_static = "1"
 native-tls = "0.2"
 openssl = "0.10"
-parse_duration = "2.0.1"
 regex = "0.2"
 serde = "1.0"
 serde_derive = "1.0"
